% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisco_vpn_class.R
\docType{data}
\name{cisco_tunnel}
\alias{cisco_tunnel}
\title{cisco_tunnel}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
cisco_tunnel
}
\description{
cisco_tunnel
}
\keyword{datasets}
