% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_unhide.R
\name{encryption}
\alias{encryption}
\alias{hide}
\alias{keychain}
\alias{unhide}
\title{Encrypt and Decrypt Private Datasets}
\usage{
hide(obj, pubkey)

unhide(hidden_obj, privatekey)

keychain()
}
\arguments{
\item{obj}{An object to hide using secure encryption}

\item{pubkey}{A public key used to encrypt data with the function \code{hide}}

\item{hidden_obj}{A hidden object to decrypt ('unhide')}

\item{privatekey}{A private key that is used to decrypt data}
}
\description{
Use the functions \code{hide} and \code{unhide} to encrypt and decrypt data.
}
\details{
\code{hide} will utilize public/private key encryption functionality provided by
the package \href{https://cran.r-project.org/web/packages/sodium/vignettes/intro.html}{sodium}.
The public key is stored locally with this package's files.
In order to decrypt the output of \code{hide}, the function \code{unhide}
will authenticate through an external deployR server. Decryption requires an interactive
environment when the user runs for the first time as credentials (username + password)
are saved locally thereafter.
}
\section{Functions}{
\itemize{
\item \code{hide}: A function to hide (encrypt) an R object

\item \code{unhide}: Use to unhide a hidden (encrypted) R object

\item \code{keychain}: used to generate the keychain if needed
}}
\examples{

# The function keychain creates a private key,
# then uses it to generate the public key.
# Both are returned in a named list.
keys <- keychain()

keys.public  <- keys$pubkey    # pubkey is used for Encryption
keys.private <- keys$privkey   # privkey is used for Decryption

## example of encryption
##
hiddenObj <- hide("hello world", keys.public)
hiddenObj

# unhide will decrypt and deserialize a hidden object
unhide(hiddenObj, keys.private)



}

