% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpu_api.R
\name{ocpu_api}
\alias{ocpu_api}
\alias{POST_OCPU}
\alias{GET_OCPU}
\alias{hpds_fn}
\alias{get_token}
\title{Authenticate to Personal R Server}
\usage{
POST_OCPU(url, pwd, ...)

GET_OCPU(url, ...)

hpds_fn(package = "rauth", fn = "check_auth", ...)

get_token(pwd = NULL, tk_name)
}
\arguments{
\item{url}{API call}

\item{pwd}{Provide password explicitly if not stored internally or as environmental variable}

\item{...}{Body arguments}

\item{package}{internal}

\item{fn}{internal}

\item{tk_name}{internal}
}
\description{
Client functions to authenticate into my personal R server
}
\section{Functions}{
\itemize{
\item \code{POST_OCPU}: POST request

\item \code{GET_OCPU}: GET request (typically comes after POST)

\item \code{hpds_fn}: tbd

\item \code{get_token}: DigitalOcean API
}}

