% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openvpn_general.R
\name{openvpn_tunnel}
\alias{openvpn_tunnel}
\title{openvpn_tunnel}
\description{
openvpn_tunnel

openvpn_tunnel
}
\section{Super class}{
\code{\link[hideR:vpn]{hideR::vpn}} -> \code{openvpn_vpn}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{openvpn_tunnel$new()}}
\item \href{#method-save}{\code{openvpn_tunnel$save()}}
\item \href{#method-disconnect}{\code{openvpn_tunnel$disconnect()}}
\item \href{#method-connect}{\code{openvpn_tunnel$connect()}}
\item \href{#method-clone}{\code{openvpn_tunnel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="hideR" data-topic="vpn" data-id="running_time">}\href{../../hideR/html/vpn.html#method-running_time}{\code{hideR::vpn$running_time()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hideR" data-topic="vpn" data-id="set_credentials">}\href{../../hideR/html/vpn.html#method-set_credentials}{\code{hideR::vpn$set_credentials()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openvpn_tunnel$new(config_path = NULL, config_file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openvpn_tunnel$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disconnect"></a>}}
\subsection{Method \code{disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openvpn_tunnel$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openvpn_tunnel$connect(
  authentification = F,
  prune_password = F,
  error_on_fail = F,
  time_out = 20,
  quiet = T,
  username = NULL,
  password = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openvpn_tunnel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
