% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\docType{class}
\name{script-class}
\alias{SCR}
\alias{script-class}
\title{A Reference Class that represents an R Script}
\description{
A Reference Class that represents an R Script
}
\section{Fields}{

\describe{
\item{\code{id}}{A character id used by rstudio api to identify script}

\item{\code{fpath}}{The encrypted full file path. Is retrieved with \code{getPath}}

\item{\code{contents}}{The encrypted contents of the script. Is retrieved with \code{getScript}}
}}
\section{Methods}{

\describe{
\item{\code{getPath(privkey = NULL)}}{Unwrap (e.g. decrypt) script path}

\item{\code{getScript(privkey = NULL)}}{Unwrap (e.g. decrypt) script}

\item{\code{wrap(pubkey = NULL)}}{Get all script information and wrap contents}
}}

