% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpn_tunnel_class.R
\docType{data}
\name{vpn_tunnel}
\alias{vpn_tunnel}
\title{vpn_tunnel}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
vpn_tunnel
}
\description{
vpn_tunnel
}
\keyword{datasets}
