% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisco_vpn_class.R
\name{cisco_tunnel}
\alias{cisco_tunnel}
\title{cisco_tunnel}
\description{
cisco_tunnel

cisco_tunnel
}
\section{Super class}{
\code{\link[hideR:vpn]{hideR::vpn}} -> \code{cisco_vpn}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cisco_tunnel$new()}}
\item \href{#method-save}{\code{cisco_tunnel$save()}}
\item \href{#method-disconnect}{\code{cisco_tunnel$disconnect()}}
\item \href{#method-connect}{\code{cisco_tunnel$connect()}}
\item \href{#method-clone}{\code{cisco_tunnel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="hideR" data-topic="vpn" data-id="running_time">}\href{../../hideR/html/vpn.html#method-running_time}{\code{hideR::vpn$running_time()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hideR" data-topic="vpn" data-id="set_credentials">}\href{../../hideR/html/vpn.html#method-set_credentials}{\code{hideR::vpn$set_credentials()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cisco_tunnel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cisco_tunnel$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disconnect"></a>}}
\subsection{Method \code{disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cisco_tunnel$disconnect(force = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cisco_tunnel$connect(prune_password = F, time_out = 20, quiet = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cisco_tunnel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
