% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openvpn_general.R
\docType{data}
\name{openvpn_tunnel}
\alias{openvpn_tunnel}
\title{openvpn_tunnel}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
openvpn_tunnel
}
\description{
openvpn_tunnel
}
\keyword{datasets}
