% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpn_tunnel_class.R
\name{vpn_tunnel}
\alias{vpn_tunnel}
\title{vpn_tunnel}
\description{
vpn_tunnel

vpn_tunnel
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{vpn_tunnel$new()}}
\item \href{#method-set_credentials}{\code{vpn_tunnel$set_credentials()}}
\item \href{#method-running_time}{\code{vpn_tunnel$running_time()}}
\item \href{#method-clone}{\code{vpn_tunnel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vpn_tunnel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_credentials"></a>}}
\subsection{Method \code{set_credentials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vpn_tunnel$set_credentials(new_file, credential_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-running_time"></a>}}
\subsection{Method \code{running_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vpn_tunnel$running_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vpn_tunnel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
